Tests: fix context re-usage, make them green when run in batch.
