+++ order/listener/src/test/kotlin/com/rarible/protocol/order/listener/service/descriptors/exchange/crypto/punks/AbstractCryptoPunkTest.kt
@@ -1,24 +1,18 @@
 package com.rarible.protocol.order.listener.service.descriptors.exchange.crypto.punks
 
-import com.ninjasquad.springmockk.MockkBean
 import com.rarible.core.common.nowMillis
 import com.rarible.core.test.wait.Wait
 import com.rarible.ethereum.common.NewKeys
 import com.rarible.ethereum.domain.EthUInt256
 import com.rarible.protocol.contracts.exchange.crypto.punks.CryptoPunksMarket
 import com.rarible.protocol.dto.ActivityDto
-import com.rarible.protocol.dto.NftOwnershipDto
-import com.rarible.protocol.nft.api.client.NftOwnershipControllerApi
 import com.rarible.protocol.order.core.configuration.OrderIndexerProperties
 import com.rarible.protocol.order.core.model.Asset
 import com.rarible.protocol.order.core.model.EthAssetType
 import com.rarible.protocol.order.core.model.OrderPriceHistoryRecord
 import com.rarible.protocol.order.core.service.PrepareTxService
-import com.rarible.protocol.order.listener.data.createNftOwnershipDto
 import com.rarible.protocol.order.listener.integration.AbstractIntegrationTest
 import io.daonomic.rpc.domain.Binary
-import io.mockk.clearMocks
-import io.mockk.coEvery
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.async
 import kotlinx.coroutines.cancelAndJoin
@@ -35,7 +29,6 @@ import org.web3j.crypto.Keys
 import org.web3j.crypto.Sign
 import org.web3j.utils.Numeric
 import reactor.core.publisher.Mono
-import reactor.kotlin.core.publisher.toMono
 import scalether.domain.Address
 import scalether.domain.request.Transaction
 import scalether.transaction.MonoGasPriceProvider
@@ -54,9 +47,6 @@ abstract class AbstractCryptoPunkTest : AbstractIntegrationTest() {
     @Autowired
     protected lateinit var prepareTxService: PrepareTxService
 
-    @MockkBean
-    private lateinit var nftOwnershipControllerApi: NftOwnershipControllerApi
-
     protected lateinit var cryptoPunksMarket: CryptoPunksMarket
 
     private lateinit var lastKafkaInstant: Instant
@@ -68,42 +58,17 @@ abstract class AbstractCryptoPunkTest : AbstractIntegrationTest() {
 
     @BeforeEach
     fun initializeCryptoPunksMarket() = runBlocking {
-        clearMocks(nftOwnershipControllerApi)
-
         cryptoPunksMarket = deployCryptoPunkMarket()
 
         // Override NFT ownership service to correctly reflect ownership of CryptoPunks.
         // By default, this service returns 1 for all ownerships, even if a punk does not belong to this address.
-        coEvery { nftOwnershipControllerApi.getNftOwnershipById(any()) } coAnswers r@{
-            val ownershipId = arg<String>(0)
-            val (tokenStr, tokenIdStr, ownerStr) = ownershipId.split(":")
-
-            val token = Address.apply(tokenStr)
-            val tokenId = EthUInt256.of(tokenIdStr)
-            val owner = Address.apply(ownerStr)
-
+        nftOwnershipAnswers = r@{ (token, tokenId, owner) ->
             if (token != cryptoPunksMarket.address()) {
-                // Return random NFT ownership with value > 0.
-                return@r createNftOwnershipDto().toMono()
+                return@r null
             }
-
-
             val realOwner = cryptoPunksMarket.punkIndexToAddress(tokenId.value).awaitSingle()
-            val value = if (owner == realOwner) BigInteger.ONE else BigInteger.ZERO
-
-            NftOwnershipDto(
-                id = ownershipId,
-                contract = token,
-                tokenId = tokenId.value,
-                owner = owner,
-                creators = emptyList(),
-                value = value,
-                lazyValue = BigInteger.ZERO,
-                date = nowMillis(),
-                pending = emptyList()
-            ).toMono()
+            if (owner == realOwner) EthUInt256.ONE else EthUInt256.ZERO
         }
-        Unit
     }
 
     protected fun checkActivityWasPublished(predicate: ActivityDto.() -> Boolean) =
